Generate a JSON parameter database for external use
To allow external programs to read and write parameters and
values on an openinverter system via CAN it is necessary to
know what parameters are available. This program runs on a
Linux system and creates a simple JSON parameter database
using a format that is similar to the serial "json" command
used by the inverter to provide similar information to the web
interface.

Compared to the serial JSON output the "value" field is
removed from each parameter as this is not relevant and the
unique CAN id is added to all parameters and values.

The main Makefile is updated to generate the new utility
"genparamdb" for Linux using the host c++ compiler.

Tests:
 - Use a new openinverter_can_tool python program to use
   the database generated by the "genparamdb" utility to list
   all available parameters.
 - Cross check the parameter and values output with the web
   interface on the same inverter
 - Modify a selection of parameters and verify unit, min and
   max limits are respected.