Fix #24 and suppress linker warning only when supported (#25)
Due to the way that gcc generates .init_array sections
containing C++ global object constructors it can result in ELF
object file sections that are marked Read Write and eXecute.
With binutils 2.39 or newer this will result in a warning. For
an embedded use case like this inverter it can safely be
suppressed.

Suppressing the warning is not possible on earlier versions of
binutils (GNU ld). The makefile now tests whether the link
option will succeed before trying to use it.

Tests:
 - Build with gcc 12.2.0 and binutils 2.39 - verify:
   - No linker warning
   - Link command contains -Wl,--no-warn-rwx-segments
 - Build with gcc 10.3.1 and binutils 2.36.1 - verify:
    - No linker warning
   - Link command does not contain -Wl,--no-warn-rwx-segments