Fix Oil pump LIN initialisation
Move the LinBus instance to the main() scope to
avoid it be deconstructed and leaving a dangling
pointer being called in Ms10Task().

Just pass a pointer of the LinBus instance to the
linM3 class rather than a pointer to the pointer. This
was causing similar dangling pointer problems
writing all over the CAN map when saving.

Tests:
 - Modify params and CAN map and save with
   ESP8266 web interface
 - Modify params and CAN map and save with
   OpenInverter CAN Tool
 - Verify params are correctly saved after reboot
 - Modify and reboot but don't save and verify
   previous values are restored