Fix variant detection for Tesla M3
The Tesla M3 board uses the PC10 pin for LIN Tx
which means it cannot be used for variant detection
using the "mini mainboard" mechanism. We do know
that it uses a STM32F103VCT6 so has 256KB of
flash. Use this fact to allow us to read the variant
resistor directly.

The resistor values read low on the V1 board so move
the values for the TESLAM3 and MG variants down to
successfully identify both. This may be down to the
noise PSU on the V1 M3DU.

Tests:
 - Boot and check reported variant on M3DU V1
 - Boot and check reported variant on Bluepill