Add 8.8kHz high-frequency resolver encmode
To allow operation on Tesla Model 3 drive units an 8.8kHz resolver encoder is
required. This is achieved with a new "HFResolver" encmode.

The exciter signal is generated using the extra TIM8 hardware on the
STM32F103VC. This is output on PC6 which is unused on the TESLAM3 hardware
variant. The exciter is synchronised in hardware to the main PWM timer TIM1.

The HFResolver position is sampled in the same way as the standard Resolver.
The reoslver is only sampled every second exciter cycle. This requires two
sample delays to pick the appropriate peak or trough in the exciter.