Refactor and simplify TeslaM3OilPump
Split out the request and response processing from
the main loop in the oil pump class. This makes the
schedule a lot more obvious.

Reduce the complexity of the polling schedule to
match the Tesla firmware. This sends requests at
100ms intervals with the status requests happing
shortly after the speed command is sent to the pump.

Tests:
 - Check with a rudimentary oil pump simulator that
   data is interchanged as before
  - Verify the contents of LIN frames with a logic
   analyser