Change automatic version number generation
To clearly differentiate between jspheubner's stm32-sine
releases and the M3_DU project use the state of the git repository
to generate the version string.

The repository's git hash and dirty state is used to prime the
checksum used by the CANopen SDO database. This should
force a download of the database ensuring consistent version
reporting in clients.

Tests:
 - Query the version with "oic read version" and verify this
   matches the output of git describe
  - Build and run unit tests