Use GitHub Action to create release firmware
This adds a release action which builds the firmware, packages
up the source and creates a draft GitHub release. It is manually
initiated and requires a pre-existing git tag to have been pushed
to the repo. The release is cerated as a draft allowing release
notes to be entered before actually publishing.

The CI build is updated to use the latest checkout action version.

Tests:
 - All to be tested after push