Mark unused inputs on Tesla M3 and MG inverters
There are several inputs which are not exposed on the Tesla
M3 and MG inverters. Ensure these report their state as "na" in
a similar way to the Rev1 and Bluepill variants.

Inputs which are not physically present but can be received via
the canio frame are not affected.

Tests:
 - Verify via "oic dumpall | grep din_" that din_mprot,
   din_emcystop and din_desat report as "na" but the rest of the
   digial inputs report the normal state.