Tighten up unit test warning level and fix warnings
Use the same warning level for test code as for the main firmware. Also,
enable the addrress and undefined sanitizers for the test code to help
catch more issues. Finally, fix various warnings in the test code.

Most of the fixes are to suppress unused parameter warnings in stub
functions, since those parameters are required to match the interface
but are not used in the stub implementations.

No problems were fixed in the main firmware.

Tests:
 - Build and run updated unit tests