Add a standard CAN map and matching DBC
ZombieVerter VCU has a standard CAN map that it expects stm32-sine
compatible inverters to be sending out. To make it easier to configure
include a CAN map JSON that OpenInverter CAN Tool can use.

The pre-existing DBC suitable for use with SavvyCAN or similar has been
extended to include all of the fields in the CAN map. Also, some missing
fields have been added to the InverterControl(0x3F) frame. Finally, the
DBC has been annotated with the source and destination for each type of
CAN frame.

Tests:
 - Verify OpenInverter CAN Tool can load the CAN map
 - Verify in SavvyCAN that the CAN map decodes the mapped fields and
   that values change as the device operates and matches values read via
   CAN SDO