Limit lowest oil temp to -30
Many newer Telsa M3 drive units do not have motor temperature sensors.
The oil temperature can act as a reasonable proxy for the motor
temperature. To avoid integer underflows with the standard CAN map we
need to limit the lowest reported temperature to -30.

This is a sensible default temperature for when the oil pump has failed.

Tests:
 - Map the oiltmp in place of tmpm in the 0x190 CAN frame
 - Use an oil pump simulator that generates temp values from -40 to 99
 - Verify the oil temp reported via CAN SDO and CAN frame 0x190 match
   and are clamped at -30
 - Unplug the oil pump and verify oiltmp is now -30 rather than 0