Add a CAN map that substitutes tmpoil for tmpm
Tesla M3 drive units without a motor sensor can use the oil temperature
as a proxy for the motor sensor. Add a CAN map than can be used to
quickly configure this.

Include in the new map in the CI firmware package.

Ensure that the firmware files are identified by git hash.

Tests:
 - Use OpenInverter CAN Tool to swap CAN maps between the
   oi-inverter-can-map.json and
   oi-inverter-can-map-no-motor-temp-sensor.json
 - Verify in SavvyCAN that the expected temperature is reported in the
   Inv_MotorTemp field and it matches that which is reported via CAN SDO
 - Verify firmware package after push