Implement a simple oil temperature controller
Create a new class OilTempController that adjusts the speed of the oil
pump based on oil temperature and operation mode.

This adds new parameters:
 - tmpoillow: Temperature below which the pump runs at nominal speed
 - tmpoilhigh - Temp above which the pump runs at maximum speed
 - pumpspeedidle - Nominal pump speed to use when not running

The existing parameter pumpspeed is used as the nominal pump speed when
running.

A full set of unit tests is provided to verify the behavior of the
controller.

Tests:
 - Run updated unit tests
 - Verify with simulated oil pump that the requested pump speed changes
   appropriately with oil temperature and operation mode.